tests pass; code functional
