+++ commands/shared.js
@@ -83,3 +83,46 @@ function getChannelsWithFlags(controller, bot, message, cb) {
     storageChannels.all(sendbackTeamChannels)
   }
 }
+
+// UNUSED
+
+function get_channel_history(channel, bot, cb) {
+  // https://github.com/howdyai/botkit/issues/840 : overwriting bot_token with app_token
+  bot.api.channels.history({token: bot.config.bot.app_token,
+                            channel: channel.id,
+                            count: 30,
+                            unreads: true},
+                            cb);
+}
+
+function channelSummary(channel, history, flags) {
+  // returns {lastFrom: ('patient'|'volunteer'),
+  //          lastTime: Date(last_message)
+  //          volunteer: <volunteer handle>
+  //         }
+  // assumes history is in reverse chronological order
+  var summary = {'id': channel.id,
+                 'name': channel.name
+                };
+  if (flags && flags.label) {
+    summary.label = flags.label
+  }
+  if (history && history.messages) {
+    for (var i=0,l=history.messages.length; i<l; i++) {
+      var h = history.messages[i];
+      if (h.subtype === 'bot_message') {
+        if (/replied/.test(h.username)) {
+          summary.lastFrom = 'volunteer'
+          summary.volunteer = h.username.replace(' replied', '')
+        } else if (h.attachments && /\/sk/.test((h.attachments[0]||{}).text || '')) {
+          summary.lastFrom = 'patient'
+        }
+        if (summary.lastFrom) {
+          summary.lastTime = new Date(Number(h.ts) * 1000)
+          return summary
+        }
+      }
+    }
+  }
+  return summary
+}
